Crypto_Tracker
README

What is this?

A crypto currency tracking app which displays the live prices of different coins. It also displays a graph which displays the historic price of a coin . The different prices at both the Indian and international exchanges would also be displayed (which has not been completely implemented).

How do I get set up?
	download contents of the repo
	open command prompt
	cd testapp
	create-react-app give_an_app_name
	npm start